convert heart rates in double

+++ lib/flutter_health_fit.dart
@@ -25,7 +25,7 @@ class HeartRateSample {
 
   HeartRateSample.fromMap(Map<String, dynamic> map)
       : dateTime = DateTime.fromMillisecondsSinceEpoch(map["timestamp"]),
-        heartRate = map["value"],
+        heartRate = map["value"].toInt(),
         metadata = map["metadata"] == null ? null : Map<String, dynamic>.from(map["metadata"]),
         sourceApp = map["metricappsource"],
         sourceDevice = map["metricdevicesource"];
